Modificaciones importantes

+++ libft/ft_memset.c
@@ -10,11 +10,11 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <stdio.h>
+#include "libft.h"
 
 void	*ft_memset(void *str, int c, size_t len)
 {
-	int		i;
+	size_t	i;
 	char	*ptr;
 
 	i = 0;
